+++ examples/graphiql-webpack/src/index.jsx
@@ -2,55 +2,18 @@ import 'regenerator-runtime/runtime.js';
 import * as React from 'react';
 import { createRoot } from 'react-dom/client';
 import { GraphiQL } from 'graphiql';
-import { useExplorerPlugin } from '@graphiql/plugin-explorer';
-import { useExporterPlugin } from '@graphiql/plugin-code-exporter';
+import { explorerPlugin } from '@graphiql/plugin-explorer';
+import { snippets } from './snippets';
+import { codeExporterPlugin } from '@graphiql/plugin-code-exporter';
 import 'graphiql/graphiql.css';
 import '@graphiql/plugin-explorer/dist/style.css';
 import '@graphiql/plugin-code-exporter/dist/style.css';
 
-const removeQueryName = query =>
-  query.replace(
-    /^[^{(]+([{(])/,
-    (_match, openingCurlyBracketsOrParenthesis) =>
-      `query ${openingCurlyBracketsOrParenthesis}`,
-  );
-
-const getQuery = (arg, spaceCount) => {
-  const { operationDataList } = arg;
-  const { query } = operationDataList[0];
-  const anonymousQuery = removeQueryName(query);
-  return (
-    ' '.repeat(spaceCount) +
-    anonymousQuery.replaceAll('\n', '\n' + ' '.repeat(spaceCount))
-  );
-};
-
-const exampleSnippetOne = {
-  name: 'Example One',
-  language: 'JavaScript',
-  codeMirrorMode: 'jsx',
-  options: [],
-  generate: arg => `export const query = graphql\`
-${getQuery(arg, 2)}
-\`
-`,
-};
-
-const exampleSnippetTwo = {
-  name: 'Example Two',
-  language: 'JavaScript',
-  codeMirrorMode: 'jsx',
-  options: [],
-  generate: arg => `import { graphql } from 'graphql'
-
-export const query = graphql\`
-${getQuery(arg, 2)}
-\`
-`,
-};
-
-const snippets = [exampleSnippetOne, exampleSnippetTwo];
-
+/**
+ * A manual fetcher implementation, you should probably
+ * just use `createGraphiQLFetcher` from `@graphiql/toolkit
+ * @returns
+ */
 const fetcher = async (graphQLParams, options) => {
   const data = await fetch(
     'https://swapi-graphql.netlify.app/.netlify/functions/index',
@@ -69,29 +32,20 @@ const fetcher = async (graphQLParams, options) => {
 };
 
 const style = { height: '100vh' };
+/**
+ * instantiate outside of the component lifecycle
+ * unless you need to pass it dynamic values from your react app,
+ * then use the `useMemo` hook
+ */
+const explorer = explorerPlugin();
+const exporter = codeExporterPlugin({ snippets });
 
 const App = () => {
-  const [query, setQuery] = React.useState('');
-  const explorerPlugin = useExplorerPlugin({
-    query,
-    onEdit: setQuery,
-  });
-  const exporterPlugin = useExporterPlugin({
-    query,
-    snippets,
-  });
-
-  const plugins = React.useMemo(
-    () => [explorerPlugin, exporterPlugin],
-    [explorerPlugin, exporterPlugin],
-  );
-
   return (
     <GraphiQL
       style={style}
-      query={query}
-      onEditQuery={setQuery}
-      plugins={plugins}
+      // eslint-disable-next-line @arthurgeron/react-usememo/require-usememo
+      plugins={[explorer, exporter]}
       fetcher={fetcher}
     />
   );